Fix labels, use time interpolation, measure convergence.
