+++ resonant-circuit/capacitor-python/run.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+set -e -u
+
+. ../../tools/log.sh
+exec > >(tee --append "$LOGFILE") 2>&1
+
+python3 capacitor.py
+
+close_log
